polinomic/requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
billiard==4.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.2
djangorestframework==3.14.0
h11==0.14.0
idna==3.10
kombu==5.4.2
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
psycopg2-binary==2.9.9
pytest==7.4.3
django_celery_beat==2.7.0
celery[redis]==5.4.0